Conversation show/hide delegate methods
Added new optional methods for `SwrveMessageDelegate` protocol:

- conversationWillBeShown:
- conversationWillBeHidden:

They provide means to react to Conversation appearing/disappearing on screen, e.g. to be able to hide keyboard when Conversation will be be shown.

Some properties and methods of SwrveConversationItemViewController are no longer public, as this class is exposed in new delegate methods now.